FunctionalTesting/PytestAgentsSDK/requirements.txt
python-dotenv
msal
msal-extensions
aiohttp
microsoft-agents-activity
microsoft-agents-copilotstudio-client

# Test dependencies
pytest
python-dotenv
pytest-html
 
# MSAL
msal
msal-extensions
 
# HTML handling
py
 
# DeepEval
deepeval
